+++ LibraryManagement/LibraryManagement.csproj
@@ -283,9 +283,11 @@
     <Compile Include="Models\BooksIssued.cs" />
     <Compile Include="Models\BooksViewModel.cs" />
     <Compile Include="Models\DBHelper.cs" />
+    <Compile Include="Models\DBResponse.cs" />
     <Compile Include="Models\IdentityModels.cs" />
     <Compile Include="Models\Images.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
+    <Compile Include="Models\Results.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repositories\BookRepositories.cs" />
     <Compile Include="Repositories\UserRepositories.cs" />
